refactor: mvc 패턴 구조로 변경
