--- src/main/java/step3/Car.java
@@ -1,17 +1,0 @@
-package step3;
-
-public class Car {
-    private int move;
-
-    public Car() {
-        this.move = 0;
-    }
-
-    public int getMove() {
-        return move;
-    }
-
-    public void go(){
-        this.move += 1;
-    }
-}
