--- src/main/java/step3/Main.java
@@ -1,16 +1,0 @@
-package step3;
-
-public class Main {
-    public static void main(String[] args) {
-        CarGame user = new CarGame();
-        int carCnt = user.ask(Message.ASK_CAR_CNT);
-        int tryCnt = user.ask(Message.ASK_TRY_CNT);
-
-        System.out.println("자동차 대수 :" + carCnt);
-        System.out.println("시도 횟수 :" + tryCnt);
-        System.out.println();
-
-        user = new CarGame(carCnt);
-        user.playGame(tryCnt);
-    }
-}
