+++ src/main/java/step3/model/Car.java
@@ -1,0 +1,14 @@
+package step3.model;
+
+public class Car {
+    private int move;
+
+    public int getMove() {
+        return move;
+    }
+
+    public int go(){
+        this.move = move+1;
+        return move;
+    }
+}
