+++ src/test/java/step3/CarGameTest.java
@@ -1,32 +1,23 @@
 package step3;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.model.Cars;
 
 import static org.assertj.core.api.Assertions.*;
 
 @DisplayName("자동차 경주 게임 테스트")
 public class CarGameTest {
-    int carCnt;
-    int playCnt;
-    CarGame game;
-
-    @BeforeEach
-    public void beforeEach() {
-        carCnt = 3;
-        playCnt = 5;
-
-        game = new CarGame(carCnt);
-    }
+    int carCnt = 3;
 
     @Test
     @DisplayName("자동차수를 입력받아 자동차 수만큼 세팅한다.")
-    public void save_cars() {
-        assertThat(game.getCarCnt()).isEqualTo(carCnt);
+    public void set_cars() {
+        Cars cars = new Cars(carCnt);
+        assertThat(cars.getCarCnt()).isEqualTo(carCnt);
 
-        for (int i=0; i<game.getCarCnt(); i++) {
-            int move = game.getCar(i).getMove();
+        for (int i=0; i<cars.getCarCnt(); i++) {
+            int move = cars.getCar(i).getMove();
             assertThat(move).isEqualTo(0);
         }
     }
@@ -39,10 +30,11 @@ public class CarGameTest {
         int carIdx = 0;
 
         //when
-        game.canMove(carIdx, randomNum);
+        Cars cars = new Cars(carCnt);
+        cars.canMove(carIdx, randomNum);
 
         //then
-        int move = game.getCar(carIdx).getMove();
+        int move = cars.getCar(carIdx).getMove();
         assertThat(move).isEqualTo(1);
     }
 
@@ -54,10 +46,11 @@ public class CarGameTest {
         int carIdx = 0;
 
         //when
-        game.canMove(carIdx, randomNum);
+        Cars cars = new Cars(carCnt);
+        cars.canMove(carIdx, randomNum);
 
         //then
-        int move = game.getCar(carIdx).getMove();
+        int move = cars.getCar(carIdx).getMove();
         assertThat(move).isEqualTo(0);
     }
 
@@ -70,12 +63,14 @@ public class CarGameTest {
         int[] randomNums    = {4, 3, 2, 8, 9};
         int[] expectedMove  = {1, 1, 1, 2, 3};
 
+        Cars cars = new Cars(carCnt);
+
         for (int i=0; i<randomNums.length; i++) {
             //when
-            game.canMove(carIdx, randomNums[i]);
+            cars.canMove(carIdx, randomNums[i]);
 
             //then
-            int move = game.getCar(carIdx).getMove();
+            int move = cars.getCar(carIdx).getMove();
             assertThat(move).isEqualTo(expectedMove[i]);
         }
     }
